@@ -1,44 +1,50 @@
 {
-  "name": "vladimir-yuldashev/flysystem-curlftp",
-  "description": "Flysystem Adapter for the FTP with cURL implementation",
-  "type": "library",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Vladimir Yuldashev",
-      "email": "misterio92@gmail.com"
+    "name": "vladimir-yuldashev/flysystem-curlftp",
+    "description": "Flysystem Adapter for the FTP with cURL implementation",
+    "type": "library",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Vladimir Yuldashev",
+            "email": "misterio92@gmail.com"
+        },
+        {
+            "name": "Alexander Rubtsov",
+            "email": "RubtsovAV@gmail.com"
+        }
+    ],
+    "require": {
+        "php": "^7.2.5|^8.0",
+        "ext-curl": "*",
+        "ext-intl": "*",
+        "league/flysystem": "^1.1"
     },
-    {
-      "name": "Alexander Rubtsov",
-      "email": "RubtsovAV@gmail.com"
-    }
-  ],
-  "require": {
-    "php": ">=5.5.9",
-    "ext-curl": "*",
-    "ext-intl": "*",
-    "league/flysystem": "^1.0"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "^4",
-    "sllh/php-cs-fixer-styleci-bridge": "^2.1",
-    "fzaninotto/faker": "^1.6",
-    "larapack/dd": "^1.1"
-  },
-  "autoload": {
-    "psr-4": {
-      "VladimirYuldashev\\Flysystem\\": "src"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "VladimirYuldashev\\Flysystem\\Tests\\": "tests"
-    }
-  },
-  "scripts": {
-    "test": "sh tests/run.sh",
-    "cs": "php-cs-fixer fix --allow-risky=yes"
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true
+    "require-dev": {
+        "phpunit/phpunit": "^8.5",
+        "larapack/dd": "^1.1",
+        "friendsofphp/php-cs-fixer": "^2.17",
+        "fakerphp/faker": "^1.12"
+    },
+    "autoload": {
+        "psr-4": {
+            "VladimirYuldashev\\Flysystem\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "VladimirYuldashev\\Flysystem\\Tests\\": "tests"
+        }
+    },
+    "scripts": {
+        "test": [
+            "@test:style",
+            "@test:unit"
+        ],
+        "test:style": "@php vendor/bin/php-cs-fixer fix --config=.php_cs.dist --allow-risky=yes --dry-run --diff --verbose",
+        "test:unit": "@php vendor/bin/phpunit",
+        "cs": "php-cs-fixer fix --allow-risky=yes",
+        "fix:style": "@php vendor/bin/php-cs-fixer fix --config=.php_cs.dist --allow-risky=yes --diff --verbose"
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
